requirements.txt
pillow==11.1.0
pycryptodome==3.22.0
